+++ kmcp/cmd/profile.go
@@ -50,31 +50,32 @@ var profileCmd = &cobra.Command{
 	Long: `Generate taxonomic profile from search results
 
 Methods:
-  1. Reference genomes can be splitted into chunks when computing
+  1. Reference genomes can be split into chunks when computing
      k-mers (sketches), which could help to increase the specificity
      via a threshold, i.e., the minimal proportion of matched chunks
      (-p/--min-chunks-fraction). (***highly recommended***)
      Another flag -d/--max-chunks-depth-stdev further reduces false positives.
-  2. We require part of the uniquely matched reads of a reference
+  2. We require a part of the uniquely matched reads of a reference
      having high similarity, i.e., with high confidence for decreasing
      the false positive rate.
   3. We also use the two-stage taxonomy assignment algorithm in MegaPath
      to reduce the false positive of ambiguous matches.
      You can also disable this step by the flag --no-amb-corr.
-     If the first stage produces thousands of candidates, you can also use
-     the flag --no-amb-corr to reduce analysis time.
+     If the stage 1/4 produces thousands of candidates, you can also use
+     the flag --no-amb-corr to reduce analysis time, which has very little
+     effect on the results
   4. Multi-aligned queries are proportionally assigned to references
      with a similar strategy in Metalign.
-  5. Input files are parsed fours times, therefore STDIN is not supported.
+  5. Input files are parsed four times, therefore STDIN is not supported.
 
 Reference:
   1. MegaPath: https://doi.org/10.1186/s12864-020-06875-6
   2. Metalign: https://doi.org/10.1186/s13059-020-02159-0
 
 Accuracy notes:
-  *. Smaller -t/--min-qcov increase sensitivity in cost of higher false
+  *. Smaller -t/--min-qcov increase sensitivity in the cost of higher false
      positive rate (-f/--max-fpr) of a query.
-  *. We require part of the uniquely matched reads of a reference
+  *. We require a part of the uniquely matched reads of a reference
      having high similarity, i.e., with high confidence for decreasing
      the false positive rate.
      E.g., -H >= 0.8 and -P >= 0.1 equals to 90th percentile >= 0.8
@@ -83,11 +84,11 @@ Accuracy notes:
      *. -P/--min-hic-ureads-prop, minimal proportion, higher values
         increase precision in cost of sensitivity.
   *. -R/--max-mismatch-err and -D/--min-dreads-prop is for determing
-     the right reference for ambiguous reads.
-  *. --keep-perfect-match is not recommended, which decreases sensitivity. 
-  *. --keep-main-match is not recommended, which affects accuracy of
+     the right reference for ambiguous reads with the algorithm in MegaPath.
+  *. --keep-perfect-matches is not recommended, which decreases sensitivity. 
+  *. --keep-main-matches is not recommended, which affects accuracy of
      abundance estimation.
-  *. -n/--keep-top-qcovs  is not recommended, which affects accuracy of
+  *. -n/--keep-top-qcovs is not recommended, which affects accuracy of
      abundance estimation.
 
 Profiling modes:
@@ -101,21 +102,17 @@ Profiling modes:
 
   Using this flag will override the relevant options.
 
-    options                      m=0    m=1   m=2   m=3    m=4   m=5
-    --------------------------   ----   ---   ---   ----   ---   ----
-    -r/--min-chunks-reads        1      20    30    50     100   100
-    -p/--min-chunks-fraction     0.2    0.5   0.7   0.8    1     1
-    -d/--max-chunks-depth-stdev  10     10    3     2      2     1.5
-    -u/--min-uniq-reads          1      20    20    20     50    50
-    -U/--min-hic-ureads          1      5     5     5      10    10
-    -H/--min-hic-ureads-qcov     0.55   0.7   0.7   0.75   0.8   0.8
-    -P/--min-hic-ureads-prop     0.01   0.1   0.2   0.1    0.1   0.15
-
-Notes on mode=0:
-  1. For detecting pathogens in samples of ultra-low depth, the flag
-     -v/--mode-0-ultra-low-depth can be used to increase the sensitivity of
-     targets with only a few reads. But note that the total reads number
-     and genome coverage will be overestimated.
+    options                       m=0    m=1   m=2   m=3    m=4   m=5
+    ---------------------------   ----   ---   ---   ----   ---   ----
+    -r/--min-chunks-reads         1      20    30    50     100   100
+    -p/--min-chunks-fraction      0.2    0.5   0.7   0.8    1     1
+    -d/--max-chunks-depth-stdev   10     10    3     2      2     1.5
+    -u/--min-uniq-reads           1      20    20    20     50    50
+    -U/--min-hic-ureads           1      5     5     5      10    10
+    -H/--min-hic-ureads-qcov      0.55   0.7   0.7   0.75   0.8   0.8
+    -P/--min-hic-ureads-prop      0.01   0.1   0.2   0.1    0.1   0.15
+    --keep-main-matches           true                            
+    --max-qcov-gap                0.4                             
 
 Taxonomy data:
   1. Mapping references IDs to TaxIds: -T/--taxid-map
@@ -125,18 +122,39 @@ Performance notes:
   1. Searching results are parsed in parallel, and the number of
      lines proceeded by a thread can be set by the flag --line-chunk-size.
   2. However using a lot of threads does not always accelerate
-     processing, 4 threads with chunk size of 500-5000 is fast enough.
-  3. If the stage 1/4 produces thousands of candidates, then stage 2/4 would
-     be very slow. You can use the flag --no-amb-corr to disable ambiguous
-     reads correction which has very little effect on the results.
+     processing, 4 threads with a chunk size of 500-5000 is fast enough.
+  3. If the stage 1/4 produces thousands of candidates, then the stage 2/4
+     would be very slow. You can use the flag --no-amb-corr to disable
+     ambiguous reads correction which has very little effect on the results.
 
 Profiling output formats:
   1. KMCP      (-o/--out-prefix)
-  2. CAMI      (-M/--metaphlan-report)
-  3. MetaPhlAn (-C/--cami-report)
+  2. CAMI      (-M/--metaphlan-report, --metaphlan-report-version, -s/--sample-id, --taxonomy-id)
+  3. MetaPhlAn (-C/--cami-report, -s/--sample-id)
+
+KMCP format:
+  Tab-delimited format with 16 columns:
+
+     1. ref,                Identifier of the reference genome
+     2. percentage,         Relative abundance of the reference
+     3. score,              The 90th percentile of qCov of uniquely matched reads
+     4. chunksFrac,         Genome chunks fraction
+     5. chunksRelDepth,     Relative depths of reference chunks
+     6. chunksRelDepthStd,  The strandard deviation of chunksRelDepth
+     7. reads,              Total number of matched reads of this reference
+     8. ureads,             Number of uniquely matched reads
+     9. hicureads,          Number of uniquely matched reads with high-confidence
+    10. refsize,            Reference size
+    11. refname,            Reference name, optional via name mapping file
+    12. taxid,              TaxId of the reference
+    13. rank,               Taxonomic rank
+    14. taxname,            Taxonomic name
+    15. taxpath,            Complete lineage
+    16. taxpathsn,          Corresponding TaxIds of taxa in the complete lineage
 
 Taxonomic binning formats:
   1. CAMI      (-B/--binning-result)
+
 `,
 	Run: func(cmd *cobra.Command, args []string) {
 		opt := getOptions(cmd)
@@ -188,8 +206,8 @@ Taxonomic binning formats:
 		maxFPR := getFlagPositiveFloat64(cmd, "max-fpr")
 		minQcov := getFlagNonNegativeFloat64(cmd, "min-query-cov")
 		topNScore := getFlagNonNegativeInt(cmd, "keep-top-qcovs")
-		keepFullMatch := getFlagBool(cmd, "keep-perfect-match")
-		keepMainMatch := getFlagBool(cmd, "keep-main-match")
+		keepFullMatch := getFlagBool(cmd, "keep-perfect-matches")
+		keepMainMatch := getFlagBool(cmd, "keep-main-matches")
 		maxScoreGap := getFlagFloat64(cmd, "max-qcov-gap")
 
 		var minReads float64
@@ -235,6 +253,9 @@ Taxonomic binning formats:
 			minHicUreads = 1
 			hicUreadsMinQcov = 0.55
 			HicUreadsMinProp = 0.1
+
+			keepMainMatch = true
+			maxScoreGap = 0.4
 		case 1:
 			minReads = 20
 			minFragsProp = 0.5
@@ -269,11 +290,6 @@ Taxonomic binning formats:
 			HicUreadsMinProp = 0.15
 		default:
 			checkError(fmt.Errorf("invalid profiling mode: %d", mode))
-		}
-
-		ultraLow := getFlagBool(cmd, "mode-0-ultra-low-depth")
-		if ultraLow && mode != 0 {
-			log.Warningf(`flag -v/--mode-0-ultra-low-depth better be used with -m 0. type "kmcp profile -h" for details`)
 		}
 
 		mode0 := mode == 0
@@ -1632,14 +1648,8 @@ Taxonomic binning formats:
 			// ----------------------
 
 			for _, c = range t.Match {
-				if ultraLow { // if mode0 {
-					if c > 0 {
-						t.FragsProp++
-					}
-				} else {
-					if c >= minReads {
-						t.FragsProp++
-					}
+				if c >= minReads {
+					t.FragsProp++
 				}
 				t.SumMatch += c
 			}
@@ -1855,15 +1865,8 @@ Taxonomic binning formats:
 										t.StatsA.Add(m.QCov)
 									}
 
-									if ultraLow { //  if mode0 {
-										t.QLen[m.FragIdx] += float64(m.QLen) / floatMsSize
-
-										t.Match[m.FragIdx] += floatOne / floatMsSize
-									} else {
-										t.QLen[m.FragIdx] += float64(m.QLen) * prop / floatMsSize
-
-										t.Match[m.FragIdx] += prop / floatMsSize
-									}
+									t.QLen[m.FragIdx] += float64(m.QLen) * prop / floatMsSize
+									t.Match[m.FragIdx] += prop / floatMsSize
 
 									if levelSpecies && theSameSpecies {
 										t.UniqMatch[m.FragIdx] += prop / floatMsSize
@@ -2077,15 +2080,8 @@ Taxonomic binning formats:
 							t.StatsA.Add(m.QCov)
 						}
 
-						if ultraLow { // if mode0 {
-							t.QLen[m.FragIdx] += float64(m.QLen) / floatMsSize
-
-							t.Match[m.FragIdx] += floatOne / floatMsSize
-						} else {
-							t.QLen[m.FragIdx] += float64(m.QLen) * prop / floatMsSize
-
-							t.Match[m.FragIdx] += prop / floatMsSize
-						}
+						t.QLen[m.FragIdx] += float64(m.QLen) * prop / floatMsSize
+						t.Match[m.FragIdx] += prop / floatMsSize
 
 						if levelSpecies && theSameSpecies {
 							t.UniqMatch[m.FragIdx] += prop / floatMsSize
@@ -2203,14 +2199,8 @@ Taxonomic binning formats:
 			// ----------------------
 
 			for _, c = range t.Match {
-				if ultraLow { // if mode0 {
-					if c > 0 {
-						t.FragsProp++
-					}
-				} else {
-					if c >= minReads {
-						t.FragsProp++
-					}
+				if c >= minReads {
+					t.FragsProp++
 				}
 				t.SumMatch += c
 			}
@@ -2573,13 +2563,13 @@ func init() {
 	profileCmd.Flags().IntP("keep-top-qcovs", "n", 0,
 		formatFlagUsage(`Keep matches with the top N qcovs for a query, 0 for all.`))
 
-	profileCmd.Flags().BoolP("keep-perfect-match", "", false,
+	profileCmd.Flags().BoolP("keep-perfect-matches", "", false,
 		formatFlagUsage(`Only keep the perfect matches (qcov == 1) if there are.`))
 
-	profileCmd.Flags().BoolP("keep-main-match", "", false,
+	profileCmd.Flags().BoolP("keep-main-matches", "", false,
 		formatFlagUsage(`Only keep main matches, abandon matches with sharply decreased qcov (> --max-qcov-gap).`))
 
-	profileCmd.Flags().Float64P("max-qcov-gap", "", 0.2,
+	profileCmd.Flags().Float64P("max-qcov-gap", "", 0.4,
 		formatFlagUsage(`Max qcov gap between adjacent matches.`))
 
 	// for matches against a reference
@@ -2665,14 +2655,12 @@ func init() {
 	profileCmd.Flags().IntP("mode", "m", 3,
 		formatFlagUsage(`Profiling mode, type "kmcp profile -h" for details. available values: 0 (for pathogen detection), 1 (higherrecall), 2 (high recall), 3 (default), 4 (high precision), 5 (higher precision).`))
 
-	profileCmd.Flags().BoolP("mode-0-ultra-low-depth", "v", false, formatFlagUsage(`Detect pathogens in samples of ultra-low depth. type "kmcp profile -h" for details.`))
-
 }
 
 // s = lambda qcov: 87.456 + 26.410*qcov - 22.008*qcov*qcov + 7.325*qcov*qcov*qcov
 func similarity(qcov float64) float64 {
 	square := qcov * qcov
-	return 87.456 + 26.410*qcov - -22.008*square + 7.325*square*qcov
+	return 87.456 + 26.410*qcov - 22.008*square + 7.325*square*qcov
 }
 
 var poolMatchResults = &sync.Pool{New: func() interface{} {